backend/requirements.txt
arrow==1.2.2
astroid==2.9.3
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.11
click==8.0.3
cryptography==36.0.1
Flask==2.0.2
h11==0.13.0
html5validator==0.4.0
idna==3.3
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
mccabe==0.6.1
outcome==1.1.0
packaging==21.3
platformdirs==2.4.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pydocstyle==6.1.1
pylint==2.12.2
pyOpenSSL==22.0.0
pyparsing==3.0.7
pytest==7.0.0
python-dateutil==2.8.2
PyYAML==6.0
requests==2.27.1
selenium==3.141.0
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
toml==0.10.2
tomli==1.2.3
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.2
wrapt==1.13.3
wsproto==1.0.0
django
discord
djangorestframework
psycopg2-binary
pyshorteners
django-cors-headers
django-environ

